+++ src/components/App/AppAccordion.vue
@@ -1,6 +1,6 @@
 <script setup>
 import { Icon } from '@iconify/vue'
-import { computed, ref, watch } from 'vue'
+import { computed, ref, watch, nextTick, useAttrs } from 'vue'
 
 const props = defineProps({
     defaultOpen: { type: Boolean, default: false },
@@ -12,20 +12,26 @@ const customGlobalClasses = computed(() => props.customGlobalClasses)
 const interaction = computed(() => props.interaction)
 const defaultOpen = computed(() => props.defaultOpen)
 const isOpen = ref(defaultOpen.value)
+const attrs = useAttrs()
 
 watch(interaction, (hasInteraction) => {
   if (hasInteraction) return
   isOpen.value = true // Set to true if interaction is enabled
 })
 
-const toggleAccordion = () => {
+const toggleAccordion = async () => {
     if (!interaction.value) return
     // Toggle the accordion open/close state
     isOpen.value = !isOpen.value
+    // If div has id attribute, scroll to it
+    if (!attrs.id) return
+    const acordionElement = document.getElementById(attrs.id)
+    await nextTick(() => acordionElement?.scrollIntoView({ behavior: 'smooth', block: 'start' }))
 }
 </script>
 <template>
   <div
+    v-bind="$attrs"
     class="accordion border-primary-200 flex-col p-4 "
     :class="{ 'border-t': isOpen, 'border-y': !isOpen, [customGlobalClasses.value]: true }"
   >